# specs/001-slide-explainer/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Slide Explainer API
  version: 0.1.0
  description: |
    Local API served by the Python backend on 127.0.0.1:8000 and proxied by Next.js at /api.
    Contract source of truth for feature 001-slide-explainer. The FastAPI app's generated
    /openapi.json MUST stay compatible with this file (contract tests enforce paths, methods,
    status codes, and schema names). Field-level semantics are defined in ../data-model.md.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /decks/{deck_id}/events:
    parameters: [{ $ref: "#/components/parameters/DeckId" }]
    get:
      tags: [runs]
      operationId: streamDeckEvents
      summary: Server-Sent Events for progress on this deck
      description: |
        Event names: run.started, run.stage, run.progress, run.paused, run.completed, run.failed,
        slide.updated, topic.updated, source.updated, document.updated, cost.updated.
        Each `data` payload is a JSON-encoded DeckEvent.
      responses:
        "200":
          description: Event stream
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/DeckEvent" }
...
  /runs/{run_id}:
    parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [runs]
      operationId: getRun
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } } }
  /runs/{run_id}/resume:
    parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [runs]
      operationId: resumeRun
      summary: Resume a failed or interrupted run from its last checkpoint (FR-030)
      responses:
        "202": { description: Resumed, content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } } }
        "409": { $ref: "#/components/responses/Problem" }
  /runs/{run_id}/cancel:
    parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [runs]
      operationId: cancelRun
      responses:
        "202": { description: Cancelling, content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } } }
...
components:
  parameters:
    DeckId:
      name: deck_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457 problem details)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            - unsupported_file
            - corrupt_file
            - too_many_slides
            - file_too_large
            - not_found
            - invalid_state
            - api_key_missing
            - api_key_invalid
            - model_not_allowed
            - provider_unavailable
            - rate_limited
            - search_unavailable
            - source_unavailable
            - libreoffice_missing
            - renderer_missing
            - validation_error
...
    StageEstimate:
      type: object
      required: [stage, model, calls, input_tokens, output_tokens, usd, path]
      properties:
        stage: { type: string }
        model: { type: string }
        calls: { type: integer }
        input_tokens: { type: integer }
        output_tokens: { type: integer }
        tool_calls: { type: integer }
        usd: { type: number }
        path: { type: string, enum: [provider, local] }
    RunEstimate:
      type: object
      required: [stages, total_usd, est_minutes]
      properties:
        stages: { type: array, items: { $ref: "#/components/schemas/StageEstimate" } }
        total_usd: { type: number }
        est_minutes: { type: number }
        budget:
          type: object
          properties:
            topics: { type: integer }
            searches_per_topic: { type: integer }
            pages_per_topic: { type: integer }
    Run:
      type: object
      required: [id, deck_id, kind, status, started_at]
      properties:
        id: { type: string }
        deck_id: { type: string }
        kind: { type: string, enum: [extract, research, process_sources, generate, qa] }
        status: { type: string, enum: [running, paused, completed, failed, cancelled] }
        stage: { type: [string, "null"] }
        progress: { type: number, minimum: 0, maximum: 1 }
        estimate: { $ref: "#/components/schemas/RunEstimate" }
        cost_usd: { type: number }
        cost_by_stage: { type: object, additionalProperties: { type: number } }
        started_at: { type: string, format: date-time }
        ended_at: { type: [string, "null"], format: date-time }
        error: { $ref: "#/components/schemas/Problem" }
    DeckEvent:
      type: object
      required: [event, deck_id, at]
      properties:
        event: { type: string }
        deck_id: { type: string }
        at: { type: string, format: date-time }
        run: { $ref: "#/components/schemas/Run" }
        slide_number: { type: integer }
        source_id: { type: string }
        document_id: { type: string }
        message: { type: string }
...